Fixed #15 - Added RawStoredProcedureManager
This manager adds raw_callproc(), which behaves the same as raw(), except that
it will execute the stored procedure and look at the first result set to create
instances of the model. The stored procedure should return the model primary
key and any other required fields. The stored procedure return value is
ignored.